Address Opus 4.8 review feedback
